Tweaks and plot reconstruction updates
Minor changes to cusize_functions to enable it to run on CSU cumulus, and expanded the plot full reconstruction function to return the binned and processed plume and height arryas.
Also added the  option single_t to plot all bin profiles of a specific time in a single plot. It doesn't make the code better to read, but very useful.